lunacon_production_restapi/requirements.txt
Django==3.0.5
django-cors-headers==3.3.0
django-crispy-forms==1.9.0
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
psycopg2-binary==2.8.5
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
virtualenv==20.0.19
wcwidth==0.1.9
websocket-client==0.57.0